requirements.txt
python-dotenv~=0.19
python-gitlab~=3.1
requests~=2.27
marshmallow~=3.14
